Cleaning up some of the internals of TIP #280.
Replaced the hashtables keyed by pointers to Proc and ByteCode
structures and providing associated information with fields in the
structures, direct pointers to the information. The change makes the
information accessible through a single pointer dereference instead of
a hashtable search.

Some information (Tcl_Obj* related) could not be inlined in this way
and likely requires a larger reworking of parser and script
structures.